+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.0] - 2019-03-11
+
+Added
+
+- Read a list of videoIds from a GCS bucket (populated by Cloud Function)
+- Cache videoId list locally
+
 ## [0.0.1] - 2019-03-05
 
 Added